#6: Rename namespace to ByteSizeLib to avoid name collision
