+++ Makefile
@@ -4,10 +4,10 @@ build:
 
 test:
 	mono tools/nuget.exe restore src/ByteSize.sln
-	xbuild src/ByteSize.Tests/ByteSize.Tests.csproj
-	mono tools/xunit/xunit.console.clr4.exe src/ByteSize.Tests/bin/Debug/ByteSize.Tests.dll
+	xbuild src/ByteSizeLib.Tests/ByteSizeLib.Tests.csproj
+	mono tools/xunit/xunit.console.clr4.exe src/ByteSizeLib.Tests/bin/Debug/ByteSizeLib.Tests.dll
 
 package:
-	xbuild src/ByteSize/ByteSize.csproj /p:Configuration=Release
-	mono tools/NuGet.exe Pack ByteSIze.nuspec
+	xbuild src/ByteSizeLib/ByteSizeLib.csproj /p:Configuration=Release
+	mono tools/NuGet.exe Pack ByteSize.nuspec
