+++ src/ByteSizeLib.Tests/ArithmeticMethods.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Xunit;
+﻿using Xunit;
 
-namespace ByteSize.Tests
+namespace ByteSizeLib.Tests
 {
     public class ArithmeticMethods
     {
