+++ src/ByteSizeLib.Tests/Properties/AssemblyInfo.cs
