+++ src/ByteSizeLib.Tests/ToStringMethod.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Xunit;
 
-namespace ByteSize.Tests
+namespace ByteSizeLib.Tests
 {
     public class ToStringMethod
     {
