+++ src/ByteSizeLib/Properties/AssemblyInfo.cs
